shop_server/requirements.txt
Django>=4.0,<5.0
gunicorn
psycopg2-binary
djangorestframework
